Update 'content/gender_feminine/02.md'
Let's link "finite verb" to a currently unmade article describing the term.

Should we delete-- ** "if the verb is active. If the verb is passive, the gender of the verb agrees with the person or the object that undergoes or receives the action (the direct object)." **
Even if the verb is passive, it still matches the subject.